+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/ChannelMessage.kt
@@ -1,11 +1,9 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.AuthorSignature
-import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
 import com.soywiz.klock.DateTime
 
@@ -18,5 +16,6 @@ data class ChannelMessage<T: MessageContent>(
     override val forwardInfo: ForwardInfo?,
     override val replyTo: Message?,
     override val replyMarkup: InlineKeyboardMarkup?,
+    override val senderBot: CommonBot?,
     val authorSignature: AuthorSignature?
-) : CommonMessage<T>
+) : CommonMessage<T>, PossiblySentViaBot